Update Dropdown component and tests, add Select component
